requirements.txt
python-decouple == 3.8
djangorestframework == 3.16.0
gunicorn == 23.0.0
django-cors-headers == 4.7.0
djangorestframework_simplejwt == 5.5.0
psycopg == 3.2.9
psycopg-binary == 3.2.9
pytest == 8.4.2
pytest-django == 4.11.1
uvicorn==0.34.0
rich==14.1.0
httpx==0.28.1
requests==2.32.5
psycopg2-binary==2.9.10